refactor(sdk): move shared communication resume engine into the SDK
The provider-agnostic snapshot-resume, task-run lookup, and out-of-band
context helpers were only usable from apps/api handlers. Queue jobs need
them too (upcoming auto-handled PR review feedback), so they move to
@roomote/sdk/server with their tests; apps/api imports through the barrel.
