+++ .docker/caddy/Caddyfile
@@ -35,7 +35,7 @@
 
 	handle @api {
 		uri strip_prefix /_roomote-api
-		reverse_proxy host.docker.internal:13001 {
+		reverse_proxy host.docker.internal:13101 {
 			lb_try_duration 10s
 			lb_try_interval 250ms
 		}
@@ -58,7 +58,7 @@
 	}
 
 	handle {
-		reverse_proxy host.docker.internal:13000 {
+		reverse_proxy host.docker.internal:13100 {
 			lb_try_duration 10s
 			lb_try_interval 250ms
 		}
