+++ packages/sdk/src/server/lib/communication/communication-snapshot-resume.ts
@@ -12,7 +12,7 @@ import {
 import {
   attachOutOfBandContextToCommunicationMessage,
   releaseCommunicationOutOfBandClaim,
-} from './communication-out-of-band-context.js';
+} from './communication-out-of-band-context';
 
 type CompletedCommunicationTaskRun = {
   id: number;
